requirements.txt
numpy 
pandas
seaborn 
matplotlib
scikit-learn
transformers==4.46.3
pandas==2.0.3
datasets
tqdm
jupyterlab
nltk
jupyterlab
nltk
torch
pyspark 
apache-airflow